requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pillow
jinja2
python-multipart
PyMuPDF
python-docx
python-pptx